# Speed up InferTypes and CInferTypes