A couple of minor improvements to performance of InferTypes and CInferTypes.

I've marked this for backporting but I did the deprecation warnings assuming it won't be included in the 1.3.1 release tomorrow.

The reason the traversal was redundant in `DefNode` was that it already mapped on its expression to get the type of the node itself. Due to declaration before use requirements, the node itself cannot be used in it's RHS.

## Benchmarks results

In a normal compilation of FIRRTL, InferTypes is run 6 times and CInferTypes is run once.

### Full compiler run

| Design | Before Runtime (s) | After Runtime (s) | Speedup |
| ---- | ---- | ---- | ----- |
| small | 47.10 ± 0.75 | 45.23 ± 0.33 | 4 % |
| medium | 117.8 ± 0.45 | 108.35 ± 2.5 | 8 % |
| large | 384.9 ± 30.7 | 380.64 ± 8.0 | uncertainty too high |

### JMH Results of LowerTypes

| Design | Before (ms) | After (ms) | Speedup |
| ---- | ---- | ---- | ---- |
| small | 283 ± 9.1 | 178 ± 12.8 | 37 % |
| medium | 698 ± 82.5 | 506 ± 23.6 | 28 % |
| large | 1715 ± 250 |  1290 ± 109 | 25% |

### Contributor Checklist

- [ ] Did you add Scaladoc to every public function/method?
- [ ] Did you update the FIRRTL spec to include every new feature/behavior?
- [ ] Did you add at least one test demonstrating the PR?
- [x] Did you delete any extraneous printlns/debugging code?
- [x] Did you specify the type of improvement?
- [x] Did you state the API impact?
- [x] Did you specify the code generation impact?
- [x] Did you request a desired merge strategy?
- [x] Did you add text to be included in the Release Notes for this change?

#### Type of Improvement

<!-- Choose one or more from the following: -->

  - performance improvement


#### API Impact

Deprecates type aliases in `InferTypes` and `CInferTypes`

#### Backend Code Generation Impact

None

#### Desired Merge Strategy

 - Rebase: You will rebase the PR onto master and it will be merged with a merge commit. 

#### Release Notes

Improve performance of InferTypes and CInferTypes

### Reviewer Checklist (only modified by reviewer)
- [ ] Did you add the appropriate labels?
- [ ] Did you mark the proper milestone (1.2.x, 1.3.0, 1.4.0) ?
- [ ] Did you review?
- [ ] Did you check whether all relevant Contributor checkboxes have been checked?
- [ ] Did you mark as `Please Merge`?