@albert-magyar 2020-05-13
> A couple of minor improvements to performance of InferTypes and CInferTypes.
> 
> I've marked this for backporting but I did the deprecation warnings assuming it won't be included in the 1.3.1 release tomorrow.
> 
> The reason the traversal was redundant in `DefNode` was that it already mapped on its expression to get the type of the node itself. Due to declaration before use requirements, the node itself cannot be used in it's RHS.

Makes sense, and several transforms will error or fail to terminate if they are run on a circuit violating this constraint, so `CheckHighForm` is already a true, hard requirement.