@jackkoenig 2020-05-13
@albert-magyar should I backport this to `1.2.x` or leave it `1.3.x`?