@albert-magyar 2020-05-13
I think 1.2.x. Both changes are just really simple and solid.